Callout views helper class, move self sizing bottom child vc inside of bottom sheet vc.
