+++ Client/Frontend/Browser/BrowserViewController.swift
@@ -1699,28 +1699,25 @@ extension BrowserViewController: TopToolbarDelegate {
     func topToolbarDidPressReload(_ topToolbar: TopToolbarView) {
         tabManager.selectedTab?.reload()
         
-        let source = SupportWebCreatorsWithTokensView()
+        let source = GetPaidForBrandedImageViewController()
+        let drawerVC = BottomSheetViewController(childViewController: source)
+
+        addChild(drawerVC)
+        view.addSubview(drawerVC.view)
+        drawerVC.view.snp.remakeConstraints {
+            $0.right.top.left.equalToSuperview()
+            $0.bottom.equalTo(view.safeArea.bottom)
+        }
         
-        let height = source.view.systemLayoutSizeFitting(UIView.layoutFittingExpandedSize).height
-        source.preferredContentSize = CGSize(width: view.frame.width, height: height)
-//        let drawerVC = BottomSheetViewController(childViewController: source)
+        
+//        let drawerVC = TranslucentBottomSheet(childViewController: source)
 //
 //        addChild(drawerVC)
 //        view.addSubview(drawerVC.view)
 //        drawerVC.view.snp.remakeConstraints {
-//            $0.right.top.left.equalToSuperview()
-//            $0.bottom.equalTo(view.safeArea.bottom).inset(footer.frame.height)
+//            $0.right.left.equalToSuperview()
+//            $0.bottom.equalTo(footer.snp.top)
 //        }
-        
-        
-        let drawerVC = TranslucentBottomSheet(childViewController: source)
-        
-        addChild(drawerVC)
-        view.addSubview(drawerVC.view)
-        drawerVC.view.snp.remakeConstraints {
-            $0.right.left.equalToSuperview()
-            $0.bottom.equalTo(footer.snp.top)
-        }
     }
     
     func topToolbarDidPressStop(_ topToolbar: TopToolbarView) {
