+++ Client/Frontend/NTP/BrandedImageCallouts/GetPaidForBrandedImageViewController.swift
@@ -8,6 +8,8 @@ import BraveShared
 
 class GetPaidForBrandedImageViewController: UIViewController {
     
+    private let viewHelper = BrandedImageCalloutHelper.CommonViews.self
+    
     let mainStackView = UIStackView().then {
         $0.axis = .vertical
         $0.translatesAutoresizingMaskIntoConstraints = false
@@ -41,23 +43,7 @@ class GetPaidForBrandedImageViewController: UIViewController {
     
     override func viewDidLoad() {
         
-        let headerStackView = UIStackView().then {
-            $0.spacing = 10
-        }
-        
-        let imageView = UIImageView(image: #imageLiteral(resourceName: "brave_rewards_button_enabled")).then {
-            $0.snp.makeConstraints {
-                $0.size.equalTo(24)
-            }
-        }
-        
-        let title = UILabel().then {
-            $0.text = "Brave Rewards"
-            $0.appearanceTextColor = .black
-            $0.font = UIFont.systemFont(ofSize: 20, weight: .semibold)
-        }
-        
-        [imageView, title].forEach(headerStackView.addArrangedSubview(_:))
+        let headerStackView = viewHelper.rewardsLogoHeader(textColor: .black, textSize: 20)
         
         headerStackView.setContentCompressionResistancePriority(UILayoutPriority(rawValue: 311), for: .vertical)
